src/main/java/ge/chalauri/transferapi/controller/AccountController.java
package ge.chalauri.transferapi.controller;

import ge.chalauri.transferapi.entity.Account;
import ge.chalauri.transferapi.service.AccountService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import static org.springframework.web.bind.annotation.RequestMethod.POST;

@RestController
@RequestMapping("/account")
public class AccountController {

    private AccountService accountService;

    @Autowired
    public AccountController(AccountService accountService) {
        this.accountService = accountService;
    }

    @RequestMapping(value = "/create", method = POST)
    public Account create(@RequestBody Account account) {
        return accountService.create(account);
    }
}
